ExamCandidate API refinement.

+++ ExamBackend/ExamBackend/src/main/java/com/company/ExamBackend/service/impl/ExamCandidateServiceImpl.java
@@ -2,12 +2,15 @@ package com.company.ExamBackend.service.impl;
 
 import com.company.ExamBackend.dto.CandidateResponseDTO;
 import com.company.ExamBackend.dto.CandidateDashboardDTO;
+import com.company.ExamBackend.exception.ExamCandidateNotFoundException;
+import com.company.ExamBackend.exception.InvalidActionException;
 import com.company.ExamBackend.mapper.CandidateMapper;
 import com.company.ExamBackend.model.ExamCandidate;
 import com.company.ExamBackend.repository.ExamCandidateRepo;
 import com.company.ExamBackend.service.ExamCandidateService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -29,8 +32,19 @@ public class ExamCandidateServiceImpl implements ExamCandidateService {
     }
 
     //Removes candidate from exam, to be implemented later.
+    @Transactional
     @Override
     public void removeCandidate(String examId, String email) {
+        ExamCandidate reqCandidate = examCandidateRepo.findByExamIdAndEmail(examId, email);
+        if(reqCandidate == null)
+        {
+            throw new ExamCandidateNotFoundException("ExamCandidate of email "+email+" not found");
+        }
+
+        if(!checkDeletionValidity(reqCandidate)) {
+            throw new InvalidActionException("Candidate of status "+reqCandidate.getStatus()+" cannot be deleted.");
+        }
+
         examCandidateRepo.deleteByExamIdAndEmail(examId, email);
     }
 
@@ -60,5 +74,16 @@ public class ExamCandidateServiceImpl implements ExamCandidateService {
 
         Instant latestPossibleFinish = now.plus(Duration.ofMinutes(durationMins));
         return !latestPossibleFinish.isAfter(end);
+    }
+
+    private boolean checkDeletionValidity(ExamCandidate candidate) {
+        return (
+                candidate.getStatus().equals("INVITED")
+                || candidate.getStatus().equals("UNINVITED")
+        );
+    }
+
+    private ExamCandidate getExamCandidate(String examId, String email) {
+        return examCandidateRepo.findByExamIdAndEmail(examId, email);
     }
 }
